refactor grideditor for py3 compatibility
