+++ mitmproxy/console/grideditor/__init__.py
@@ -1,0 +1,2 @@
+from .editors import *  # noqa
+from . import base  # noqa
